requirements.txt
httpx>=0.16.1
matplotlib>=3.2.0
pygtrie>=2.0
pytz>=2021.1
numpy>=1.20.1
aiohttp>=3.7.3
loguru>=0.5.3
lxml>=4.6.2
requests>=2.25.1
ujson>=4.0.2
fuzzywuzzy>=0.18.0
peony_twitter[all]~=1.1.7
zhconv>=1.4.1
beautifulsoup4>=4.9.3
dataclasses>=0.8
nonebot2==2.0.0a11
nonebot_plugin_apscheduler==0.1.2
peony>=1.0.34
Pillow>=8.1.0